requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
requests>=2.31.0
reportlab>=4.0.0
Pillow>=10.0.0
python-pptx>=0.6.21
fastapi
uvicorn
google-generativeai